requirements.txt
pytesseract
selenium
pillow
loguru